refactor(python): Move metadata utils to `meta` module
