+++ py-polars/polars/utils/__init__.py
@@ -4,7 +4,6 @@ Utility functions.
 Functions that are part of the public API are re-exported here.
 """
 from polars.utils._scan import _execute_from_rust
-from polars.utils.build_info import build_info
 from polars.utils.convert import (
     _date_to_pl_date,
     _datetime_for_any_value,
@@ -17,18 +16,12 @@ from polars.utils.convert import (
     _to_python_time,
     _to_python_timedelta,
 )
-from polars.utils.meta import get_index_type, threadpool_size
-from polars.utils.show_versions import show_versions
 from polars.utils.various import NoDefault, _polars_warn, is_column, no_default
 
 __all__ = [
     "NoDefault",
-    "build_info",
-    "get_index_type",
     "is_column",
     "no_default",
-    "show_versions",
-    "threadpool_size",
     # Required for Rust bindings
     "_date_to_pl_date",
     "_datetime_for_any_value",
